Create AOOFUNDING.yml
There are two primary development components to the Array of Opportunity; property development and business operation development.  We want to allow participants the opportunity for equity in exchange for their investment and compensation for lasting superior value adds to our work.  To that end we are establishing a fund to invest in aspects of development where needed.